% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi.R
\name{midi_play}
\alias{midi_play}
\alias{midi_convert}
\alias{midi_data}
\title{Play or convert a midi file}
\usage{
midi_play(
  midi = demo_midi(),
  soundfont = soundfont_path(),
  audio.driver = NULL,
  settings = list(),
  verbose = interactive()
)

midi_convert(
  midi = demo_midi(),
  soundfont = soundfont_path(),
  output = "output.mp3",
  settings = list(),
  verbose = interactive()
)

midi_data(midi = demo_midi(), verbose = interactive())
}
\arguments{
\item{midi}{path to the midi file}

\item{soundfont}{path to the soundfont}

\item{audio.driver}{which audio driver to use,
see \href{https://www.fluidsynth.org/api/CreatingAudioDriver.html}{fluidsynth docs}}

\item{settings}{a named vector with additional settings from \code{\link[=fluidsynth_setting_list]{fluidsynth_setting_list()}}}

\item{verbose}{print some progress status to the terminal}

\item{output}{filename of the output. The out}
}
\description{
Renders a midi file to your audio device or to a wav file. Additional settings
can be specified, see \link{fluidsynth_setting_list} for availble options.
On Linux you may need to specify an \code{audio.driver} that works for your hardware.
}
\details{
The \code{midi_convert} function internally uses fluidsynth to generate a raw wav file,
and then \code{\link[av:encoding]{av::av_audio_convert()}} to convert into the requested about format. See
\code{\link[av:formats]{av::av_muxers()}} for supported output formats and their corresponding file extension.

You need a soundfont to play midi.
A free soundfont called \href{https://www.schristiancollins.com/generaluser.php}{GeneralUser-GS} is
included with this package.
}
\examples{
midi_convert(settings = list('synth.sample-rate'= 22050), output =  'lowquality.mp3')
unlink("lowquality.mp3")
}
