% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{fluidsynth_settings}
\alias{fluidsynth_settings}
\alias{fluidsynth_setting_list}
\alias{fluidsynth_setting_options}
\alias{fluidsynth_setting_default}
\title{Fluidsynth settings}
\usage{
fluidsynth_setting_list()

fluidsynth_setting_options(setting)

fluidsynth_setting_default(setting)
}
\arguments{
\item{setting}{string with one of the options listed in \code{\link[=fluidsynth_setting_list]{fluidsynth_setting_list()}}}
}
\description{
Get available settings and their types.
}
\examples{
# List available settings:
fluidsynth_setting_list()
fluidsynth_setting_options('audio.driver')
fluidsynth_setting_default('synth.sample-rate')
}
