% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi.R
\name{midi_play}
\alias{midi_play}
\title{Play a midi file}
\usage{
midi_play(midi = demo_midi(), soundfont = general_user_gs())
}
\arguments{
\item{midi}{path to the midi file}

\item{soundfont}{path to the soundfont}
}
\description{
You need a soundfont to play midi.
A free soundfont called \href{https://www.schristiancollins.com/generaluser.php}{GeneralUser-GS} is
included with this package.
}
