% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundfonts.R
\name{soundfonts}
\alias{soundfonts}
\alias{soundfont_path}
\alias{soundfont_generaluser_gs}
\title{Managing soundfonts}
\usage{
soundfont_path(download = TRUE)

soundfont_generaluser_gs(download = TRUE)
}
\arguments{
\item{download}{automatically download soundfont if none exists.}
}
\description{
FluidSynth requires a soundfont to synthesize a midi. On Linux distributions
some soundfounds are often preinstalled, though their quality varies. If your
midi sounds very poor, try seleting another soundfont.
}
\details{
\href{https://schristiancollins.com/generaluser}{GenralUser-GS} by S. Christian Collins
is a nice free soundfont. The \code{soundfont_generaluser_gs()} function will
automatically download a copy to for this package.
}
