% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi.R
\name{midi_play}
\alias{midi_play}
\alias{midi_convert}
\title{Play a midi file}
\usage{
midi_play(
  midi = demo_midi(),
  soundfont = general_user_gs(),
  audio.driver = NULL,
  progress = TRUE
)

midi_convert(
  midi = demo_midi(),
  soundfont = general_user_gs(),
  output = "output.wav",
  progress = TRUE
)
}
\arguments{
\item{midi}{path to the midi file}

\item{soundfont}{path to the soundfont}

\item{audio.driver}{which audio driver to use,
see \href{https://www.fluidsynth.org/api/CreatingAudioDriver.html}{fluidsynth docs}}

\item{progress}{print status progress to the terminal}

\item{output}{filename of the output. It is recommended to use wav output.}
}
\description{
You need a soundfont to play midi.
A free soundfont called \href{https://www.schristiancollins.com/generaluser.php}{GeneralUser-GS} is
included with this package.
}
