% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundfonts.R
\name{soundfonts}
\alias{soundfonts}
\alias{soundfont_path}
\alias{soundfont_download}
\title{Managing soundfonts}
\usage{
soundfont_path(download = FALSE)

soundfont_download()
}
\arguments{
\item{download}{automatically download soundfont if none exists.}
}
\value{
the path to a local soundfont to synthesize a midi file.
}
\description{
FluidSynth requires a soundfont to synthesize a midi. On Linux distributions
some soundfonts are often preinstalled, though their quality varies. If your
midi sounds very poor, try using another soundfont.
}
\details{
\href{https://schristiancollins.com/generaluser}{GeneralUser-GS} by S. Christian Collins
is a nice free soundfont. You can use \code{soundfont_download()} to install a copy
of this soundbank for use by this package.
}
\seealso{
Other fluidsynth: 
\code{\link{fluidsynth_settings}},
\code{\link{midi_play}()}
}
\concept{fluidsynth}
